Update tests to account for new data
Okay - my bad.  How could adding additional locations and bookables break unit tests.  No way, so I didn't run tests locally.  I am a bad man.  Also, I will have to think of a better was of testing this functionality so that the data setup is local to the test construct and now reliant on the existance of ... production data values.